refactor(lib): move logger
