+++ app/authentication.js
@@ -1,6 +1,6 @@
 const jwt = require('express-jwt')
 const jwksRsa = require('jwks-rsa')
-const logger = require('../lib/logger.js')
+const logger = require('./lib/logger.js')
 
 const secret = jwksRsa.expressJwtSecret({
   cache: true,
