+++ app/lib/logger.js
