+++ app/resources/services/integrations/coil.js
@@ -5,7 +5,7 @@ const InternalOAuthError = require('passport-oauth2').InternalOAuthError
 const axios = require('axios')
 const { User } = require('../../../db/models')
 const appURL = require('../../../lib/url.js')
-const logger = require('../../../../lib/logger.js')
+const logger = require('../../../lib/logger.js')
 
 const {
   findUser,
