+++ app/resources/v1/__tests__/user.test.js
@@ -4,7 +4,7 @@ import { setupMockServer } from '../../../../test/helpers/setup-mock-server'
 import user from '../users'
 
 jest.mock('../../../db/models')
-jest.mock('../../../../lib/logger')
+jest.mock('../../../lib/logger')
 
 // mockUser is setting a mock 'sub' (which is oAuth shorthand for 'subject'),
 // so that the below tests are mock authenticated. Don't confuse this with actual user data
