+++ app/resources/v1/streets.js
@@ -1,7 +1,7 @@
 const { v1: uuidv1 } = require('uuid')
 const { isArray } = require('lodash')
 const { ERRORS, asStreetJson } = require('../../../lib/util')
-const logger = require('../../../lib/logger.js')
+const logger = require('../../lib/logger.js')
 const { User, Street, Sequence } = require('../../db/models')
 
 exports.post = async function (req, res) {
