+++ app/resources/v1/user_session.js
@@ -1,4 +1,4 @@
-const logger = require('../../../lib/logger.js')
+const logger = require('../../lib/logger.js')
 
 exports.delete = async function (req, res) {
   // In order to sign out, make sure the user's session cookies
