+++ lib/request_handlers/metatags.js
@@ -1,6 +1,6 @@
 const request = require('request')
 const { User, Street } = require('../../app/db/models')
-const logger = require('../../lib/logger.js')
+const logger = require('../../app/lib/logger.js')
 const appURL = require('../../app/lib/url')
 
 const ANON_CREATOR = '-'
