+++ lib/request_handlers/request_log.js
@@ -1,4 +1,4 @@
-const logger = require('../../lib/logger.js')
+const logger = require('../../app/lib/logger.js')
 
 module.exports = function (req, res, next) {
   const contentType = req.headers['content-type'] || ''
